IXGraphBase changes
Mark as final to prevent inheritance.
Change to using SimpleBiMap as opposed to boost::bimap for bidirectional maps
Make vertex/edge label a union.
Remove usage of IXBool. All functions now assume that they have been provided valid parameters. It is up to the caller to ensure that this is the case.
Completed documentation of class