Bond Order Integration
Add indigo-bondorder repo as a submodule. The namespace has been changed from indigox to indigo-bondorder, so hopefully it is now straightforward to integrate the formal charge and bond order calculation.

Note that this submodule points to a specific branch, seb/change-namespace, and points to my personal fork of the indigo-bondorder repo (for now). origin has been set to cuboxylate/indigo-bondorder. If this works I will reintegrate into the full repo and redirect this submodule