Bond Order Integration
I have it working! Been stuck on an OOME running a med-sized molecule through bondorder on A* or FTP, but turning on ADD_EDGES_TO_TD worked to get around it.

Committing changes to push some BO options to user in CP, and well as finish the basic integration. Cleaned up my example classes too and made a general case one.